requirements.txt
appdirs==1.4.4
asgiref==3.3.4
book-me==0.1.0
CacheControl==0.12.6
cachy==0.3.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cleo==0.8.1
clikit==0.6.2
crashtest==0.3.1
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.4
django-allauth==0.44.0
djangorestframework==3.12.4
filelock==3.0.12
gunicorn==20.1.0
html5lib==1.1
idna==2.10
keyring==21.8.0
lockfile==0.12.2
msgpack==1.0.2
oauthlib==3.1.1
packaging==20.9
pastel==0.2.1
pexpect==4.8.0
pkginfo==1.7.0
poetry==1.1.7
poetry-core==1.0.3
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pycparser==2.20
PyJWT==2.1.0
pylev==1.4.0
pyparsing==2.4.7
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
shellingham==1.4.0
six==1.16.0
sqlparse==0.4.1
tomlkit==0.7.2
urllib3==1.26.6
virtualenv==20.4.7
webencodings==0.5.1
whitenoise==5.2.0
